+++ Core/SoarKernel/src/semantic_memory/smem_instance.cpp
@@ -177,7 +177,7 @@ Symbol* SMem_Manager::get_current_iSTI_for_LTI(uint64_t pLTI_ID, goal_stack_leve
 {
     id_to_sym_map::iterator lIter;
 
-    dprint(DT_SMEM_INSTANCE, "Getting current iSTI for lti ID %u at level %d.\n", pLTI_ID, pLevel);
+    dprint(DT_SMEM_INSTANCE, "Getting current iSTI for lti ID %u at level %d.\n", pLTI_ID, static_cast<int64_t>(pLevel));
     lIter = lti_to_sti_map.find(pLTI_ID);
 
     if (lIter != lti_to_sti_map.end())
